requirements.txt
numpy==1.22.3
pandas==1.4.1
streamlit==1.7.0
easyocr==1.4.2
opencv-python-headless==4.5.4.60
Pillow==8.2.0
torch==1.11.0
torchvision==0.12.0
click==8.0.4
PyYAML==5.3.1
psutil==5.7.0
requests==2.26.0
regex==2020.6.8
beautifulsoup4==4.9.1
tensorflow==2.8.0
tensorflow-data-validation==1.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.7.0
tensorflow-model-analysis==0.38.0
tensorflow-serving-api==2.8.0
tensorflow-transform==1.7.0
keras==2.8.0
keras-adaptive-softmax==0.9.0
Keras-Applications==1.0.8
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-ocr==0.8.9
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.51.0
keras-trans-mask==0.6.0
keras-transformer==0.40.0
keras-transformer-xl==0.14.0
keras-tuner==1.1.2
keras-xlnet==0.20.0
transformers==4.17.0
plotly==5.6.0

